refactor: 채팅 받기 리팩토링
채널 바깥에서도 채팅을 받을 수 있어야 하기 때문에 SocketLayer에서 채팅 받을 수 있도록 변경, 스크롤바를 전역 상태로 설정해서 채팅시 스크롤바 조절
